Pulled object processing out into the new architecture
